mypro2/requirements.txt
api==0.0.7
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
Django==4.2.20
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.10
exceptiongroup==1.2.2
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==2.1.5
nose==1.3.7
packaging==25.0
pluggy==1.5.0
pycparser==2.22
PyMySQL==1.1.1
pytest==8.3.5
pytest-django==4.11.1
pytest-html==4.1.1
pytest-metadata==3.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3